+++ assets/css/admin.css
@@ -1,1 +1,1 @@
-#sftExt_color_picker_btn_bg_color,#sftExt_color_picker_btn_font_color{padding:10px 20px;border:1px solid #ddd;-webkit-box-shadow:inset 0 1px 2px rgba(0,0,0,.07);box-shadow:inset 0 1px 2px rgba(0,0,0,.07);display:inline-block;color:#333;text-align:center;cursor:pointer}.sftExt_alert{color:#000000;background:#FFCCCC}.sftExt_alert th{padding-left:10px}
+#sftExt_color_picker_btn_bg_color,#sftExt_color_picker_btn_font_color{padding:10px 20px;border:1px solid #ddd;-webkit-box-shadow:inset 0 1px 2px rgba(0,0,0,.07);box-shadow:inset 0 1px 2px #00000012;display:inline-block;color:#333;text-align:center;cursor:pointer}.sftExt_alert{color:#000;background:#FFCCCC}.sftExt_alert th{padding-left:10px}
