+++ assets/css/frontend.css
@@ -1,1 +1,1 @@
-#sftExt-frontend-button{position:fixed;border:none;display:inline-block;cursor:pointer;z-index:9999;left:20px;top:20px;padding:10px 20px;font-family:inherit;width:25px;height:25px}@media screen,speech{#sftExt-frontend-button .sr-only{position:absolute;left:-9999px;width:1px;height:1px;overflow:hidden}}#sftExt-frontend-button.round{border-radius:50%;padding:10px;text-align:center;width:45px;height:45px}#sftExt-frontend-button.round>i{position:relative;display:block;font-size:25px;top:50%;-webkit-transform:translateY(-50%);-moz-transform:translateY(-50%);-ms-transform:translateY(-50%);-o-transform:translateY(-50%);transform:translateY(-50%)}#sftExt-frontend-button.square{border-radius:0;padding:10px;text-align:center;width:45px;height:45px}#sftExt-frontend-button.square>i{position:relative;display:block;font-size:25px;top:50%;-webkit-transform:translateY(-50%);-moz-transform:translateY(-50%);-ms-transform:translateY(-50%);-o-transform:translateY(-50%);transform:translateY(-50%)}#sftExt-frontend-button.rectangle{padding:10px 20px;height:auto;width:auto;display:-webkit-inline-box;display:-webkit-inline-flex;display:-moz-inline-box;display:inline-flex;-webkit-box-orient:vertical;-webkit-box-direction:normal;-webkit-flex-direction:column;-moz-box-orient:vertical;-moz-box-direction:normal;flex-direction:column}#sftExt-frontend-button.rectangle:before,#sftExt-frontend-button.rectangle:after{content:'';-webkit-box-flex:1;-webkit-flex:1;-moz-box-flex:1;flex:1}#sftExt-frontend-button.rectangle .sftExt-inner{position:relative;text-align:center}#sftExt-frontend-button.rectangle .sftExt-inner>i{margin-right:10px}#sftExt-frontend-button.rectangle .sftExt-inner:before{position:absolute;left:0;top:0;height:20px;width:20px}#sftExt-frontend-button.bottom{top:auto;bottom:20px}#sftExt-frontend-button.top{top:20px}#sftExt-frontend-button.right{left:auto;right:20px}#sftExt-frontend-button.left{left:20px}
+#sftExt-frontend-button{position:fixed;border:none;display:inline-block;cursor:pointer;z-index:9999;left:20px;top:20px;padding:10px 20px;font-family:inherit;width:25px;height:25px}@media screen,speech{#sftExt-frontend-button .sr-only{position:absolute;left:-9999px;width:1px;height:1px;overflow:hidden}}#sftExt-frontend-button.round{border-radius:50%;padding:10px;text-align:center;width:45px;height:45px}#sftExt-frontend-button.round>i{position:relative;display:block;font-size:25px;top:50%;-webkit-transform:translateY(-50%);-moz-transform:translateY(-50%);-ms-transform:translateY(-50%);transform:translateY(-50%)}#sftExt-frontend-button.square{border-radius:0;padding:10px;text-align:center;width:45px;height:45px}#sftExt-frontend-button.square>i{position:relative;display:block;font-size:25px;top:50%;-webkit-transform:translateY(-50%);-moz-transform:translateY(-50%);-ms-transform:translateY(-50%);transform:translateY(-50%)}#sftExt-frontend-button.rectangle{padding:10px 20px;height:auto;width:auto;display:-webkit-inline-box;display:-webkit-inline-flex;display:-moz-inline-box;display:inline-flex;-webkit-box-orient:vertical;-webkit-box-direction:normal;-webkit-flex-direction:column;-moz-box-orient:vertical;-moz-box-direction:normal;flex-direction:column}#sftExt-frontend-button.rectangle:before,#sftExt-frontend-button.rectangle:after{content:"";-webkit-box-flex:1;-webkit-flex:1;-moz-box-flex:1;flex:1}#sftExt-frontend-button.rectangle .sftExt-inner{position:relative;text-align:center}#sftExt-frontend-button.rectangle .sftExt-inner>i{margin-right:10px}#sftExt-frontend-button.rectangle .sftExt-inner:before{position:absolute;left:0;top:0;height:20px;width:20px}#sftExt-frontend-button.bottom{top:auto;bottom:20px}#sftExt-frontend-button.top{top:20px}#sftExt-frontend-button.right{left:auto;right:20px}#sftExt-frontend-button.left{left:20px}
