+++ package.json
@@ -4,17 +4,20 @@
 	"description": "",
 	"main": "index.js",
 	"scripts": {
-		"dev": "vite",
+        "dev": "vite",
+        "build": "vite build",
 		"build:frontend": "postcss ./src/css/frontend.css -o assets/css/frontend.css",
 		"build:admin": "postcss ./src/css/admin.css -o assets/css/admin.css"
 	},
 	"author": "",
 	"license": "ISC",
 	"devDependencies": {
-		"autoprefixer": "^10.4.14",
-		"postcss": "^8.4.24",
+		"autoprefixer": "^10.4.15",
+		"cssnano": "^6.0.1",
+		"postcss": "^8.4.29",
+		"postcss-cli": "^10.1.0",
 		"postcss-minify": "^1.1.0",
 		"postcss-preset-env": "^9.0.0",
-		"vite": "^4.3.9"
+		"vite": "^4.4.9"
 	}
 }
